Move level editor into own file, start creating interface for selecting what you want to place
