+++ public/index.html
@@ -15,6 +15,8 @@
       <div id="pause-menu">
         <h2>PAUSED</h2>
       </div>
+      <div id="toolbox">
+      </div>
     </div>
 
     <div id="help-text">
